requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
distlib==0.3.1
Django==3.0.8
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-filter==2.3.0
filelock==3.0.12
kademlia==2.2.1
pipenv==2020.8.13
pytz==2020.1
rpcudp==4.0.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
u-msgpack-python==2.7.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
